+++ packages/components/src/index.ts
@@ -9,3 +9,4 @@ export * from './components/Map';
 export * from './components/PdfViewer';
 export * from "./components/Excel";
 export * from "./components/BucketViewer";
+export * from "./components/Iframe";